Update VSCode settings and add new functions and methods
Add "files.readonlyInclude" setting to include TypeScript files in read-only mode.

Add new function "inspect" to inspect JavaScript objects.

Add new function "log" to log messages to the console.

Add new classes "FixedTokens", "StopToken", "ConstrainedToken", and "AiciAsync".

Add new methods and constructor to the "Constraint" class.

Add new methods and constructor to the "NextToken" class.

Add new methods and constructor to the "AiciAsync" class.

Add new methods "fixed" and "check_var" to the "AiciAsync" class.

Add new type aliases "SeqId", "int", and "Buffer".

Update the "pre_process" method in the "AiciAsync" class to log the current token.

Update the "check_var" function to throw an error if the variable is unset.

Update the "gen_text" function to return a Promise.

Update the "AiciAsync" constructor to bind the callback methods.

Update the "AiciAsync" constructor to set the global callback object.

Update the "AiciAsync" constructor to call the initialization function.

Update the "AiciAsync" constructor to check if the prompt is available.

Update the "AiciAsync" constructor to set the prompt callback.

Update the "AiciAsync" constructor to set the pre-process callback.

Update the "AiciAsync" constructor to set the mid-process callback.

Update the "AiciAsync" constructor to set the post-process callback.

Update the "AiciAsync" constructor to set the stop callback.

Update the "AiciAsync" constructor to set the error callback.

Update the "AiciAsync" constructor to set the log callback.

Update the "AiciAsync" constructor to set the info callback.

Update the "AiciAsync" constructor to set the warn callback.

Update the "AiciAsync" constructor to set the debug callback.

Update the "AiciAsync" constructor to set the trace callback.